refactor: introduce app factory
- move legacy chat logic into fred_zulip_bot.legacy_app for reuse
- add shared pydantic models plus app factory and route modules with health endpoints
- expose create_app() from main.py so uvicorn can use the factory
