+++ fred_zulip_bot/apps/__init__.py
@@ -1,0 +1,1 @@
+"""Application entrypoints."""
