+++ fred_zulip_bot/apps/api/__init__.py
@@ -1,0 +1,1 @@
+"""FastAPI app factory package."""
