+++ fred_zulip_bot/apps/api/routes/__init__.py
@@ -1,0 +1,1 @@
+"""API route registrations."""
