+++ fred_zulip_bot/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core utilities and shared models."""
