+++ main.py
@@ -1,364 +1,12 @@
-from fastapi import FastAPI, BackgroundTasks, HTTPException
-from fastapi.responses import JSONResponse
-from pydantic import BaseModel
-from fastapi.middleware.cors import CORSMiddleware
-from config import config
-from typing import Any
-import google.generativeai as genai
-import mysql.connector
-import requests
-import re
-import os
-import json
-from logger import logger
+"""Main entrypoint exposing the FastAPI application factory."""
 
+from __future__ import annotations
 
-if config.TEST_MODE:
-    zulip_bot_email = config.ZULIP_BOT_EMAIL_TEST
-    zulip_bot_token = config.ZULIP_BOT_TOKEN_TEST
-    zulip_auth_token = config.ZULIP_AUTH_TOKEN_TEST
-else:
-    zulip_bot_email = config.ZULIP_BOT_EMAIL
-    zulip_bot_token = config.ZULIP_BOT_TOKEN
-    zulip_auth_token = config.ZULIP_AUTH_TOKEN
+from fred_zulip_bot.apps.api.app import create_app
 
-# Sends a message to Zulip (either private or stream)
-def send_zulip_message(to, msg_type, subject, content, channel_name):
-    data = {
-        "type": msg_type,
-        "to": to,
-        "content": content
-    }
-    if msg_type == "stream": # Stream messages require subject + stream name
-        data["subject"] = subject
-        data["to"] = channel_name
+app = create_app()
 
-    requests.post(
-        f"{config.ZULIP_SITE}/api/v1/messages",
-        data=data,
-        auth=(zulip_bot_email, zulip_bot_token)
-    )
+if __name__ == "__main__":
+    import uvicorn
 
-# Executes a SQL query against the database and returns results
-def submit_query(query):
-    try:
-        conn = mysql.connector.connect(
-            host = config.DB_HOST,
-            port = 3306,
-            database = config.DB_NAME,
-            user = config.DB_USER,
-            password = config.DB_PASSWORD,
-            charset = 'utf8mb4',
-            collation = 'utf8mb4_unicode_ci'
-        )
-
-        db = conn.cursor()
-        db.execute(query)
-        rows = db.fetchall()
-        result = ""
-        for row in rows:
-            result += str(row) + ", "
-
-        db.close()
-        conn.close()
-        return result
-    except Exception as e:
-        # Returns value for bot to salvage message on error
-        logger.error("invalid sql generated", exc_info=True)
-        return "salvage"
-
-# Disallowed SQL keywords to prevent unsafe operations
-FORBIDDEN_SQL_KEYWORDS = [
-    "INSERT", "UPDATE", "DELETE", "DROP", "TRUNCATE", "ALTER", "REPLACE", "CREATE"
-]
-
-# Checks if SQL query is safe (only read operations allowed)
-def is_safe_sql(query: str) -> bool:
-    cleaned_query = re.sub(r"\s+", " ", query).strip().upper()
-
-    for keyword in FORBIDDEN_SQL_KEYWORDS:
-        if re.search(rf"\b{keyword}\b", cleaned_query):
-            return False
-
-    return True
-
-
-# Load database schema (DDL statements) to provide context for AI SQL generation
-with open("context/DDLs.rtf", "r") as f:
-    database_context = f.read()
-
-# Load additional rules for how SQL should be generated
-with open("context/system_prompt_rules.txt") as f:
-    system_prompt_rules = f.read()
-
-# Initialize FastAPI app (disables auto docs for security)
-app = FastAPI(docs_url=None, redoc_url=None, openapi_url=None)
-
-# Enable CORS if you're calling from a browser/frontend
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["*"],
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-
-# Prompt used to classify user intent
-intent_prompt = (
-    "You are an intent classifier. A user has sent a message."
-    "Your task is to classify their intent into one of the following categories:\n"
-    "- database: The user wants to query or access information from the database.\n"
-    "- chatbot: The user is asking about the chatbot itself, like its purpose, name, capabilities, etc.\n"
-    "- other: Anything that does not fit the above categories.\n"
-    "Respond ONLY with one of the following words: database, chatbot, other.\n"
-)
-
-# Prompt for chatbot-related questions (about Fred itself)
-chatbot_prompt = (
-    "You are Fred, an AI-powered assistant integrated with Zulip. The database you are an interface for has a lot"
-    "of information relating to unfoldingWord’s work in Bible translation. Fred generates safe, read-only SQL"
-    "queries based on the user's natural language request. However, Fred is not an SQL assistant and doesn’t help"
-    "the user with questions about SQL. It has access to the database schema and follows strict system rules when"
-    "generating SQL. It executes safe queries (only SELECT/read operations) and summarizes the results in natural"
-    "language. The user is asking about you directly—a question like what you do, how you work, or what your name is."
-    "Answer clearly and briefly, as a helpful assistant would. Don't generate SQL or refer to specific database contents."
-)
-
-# Prompt for unsupported questions
-other_prompt = (
-    "You are Fred, an AI-powered assistant integrated with Zulip. The database you are an interface for has a lot"
-    "of information relating to unfoldingWord’s work in Bible translation. Fred generates safe, read-only SQL"
-    "queries based on the user's natural language request. However, Fred is not an SQL assistant and doesn’t help"
-    "the user with questions about SQL. It has access to the database schema and follows strict system rules when"
-    "generating SQL. It executes safe queries (only SELECT/read operations) and summarizes the results in natural"
-    "language. The user has asked something of you that is an unsupported function of this chatbot. Kindly explain"
-    "to the user that you can't help them with that, and redirect them by informing them of things you can do."
-)
-
-# Prompt for SQL generation — strict rules enforced
-sql_prompt = (
-    "You are an SQL assistant.You will generate SQL queries based on the the user's request and the database information that was given to you."
-    "Only return the SQL query — no explanation, no Markdown, no code block formatting."
-    "You are a read-only assistant. Under no circumstances should you ever modify the database."
-    "If the user asks you to do so, inform them that you are not able to do that. \n"
-    f"Here is the database schema: \n{database_context}"
-    f"Here are rules you must adhere to when creating sql queries: \n{system_prompt_rules}"
-
-)
-
-# Prompt for summarizing SQL results into natural language
-answer_prompt = (
-    "You are a data summarizer. The user asked a question and you've been given the raw SQL result."
-    "Based on that result, write a clear and concise natural-language answer."
-    "Make sure to restate the user's question in the answer. If asked about Population, cite the"
-    "datasource either as Joshua Project or Progress Bible based on the SQL table used."
-)
-
-# Initialize genai client
-genai.configure(api_key=config.GENAI_API_KEY)
-
-current_model = "gemini-2.5-pro"
-
-# Pydantic model for Zulip message
-class ZulipMessage(BaseModel):
-    content: str
-    display_recipient: Any
-    sender_email: str
-    subject: str
-    type: str
-
-# Incoming request wrapper (includes Zulip message + auth token)
-class ChatRequest(BaseModel):
-    message: ZulipMessage
-    token: str
-
-# Outgoing response structure
-class ChatResponse(BaseModel):
-    response: str
-
-# Store chat history per user
-CHAT_HISTORY_DIR = "./data/chat_histories"
-os.makedirs(CHAT_HISTORY_DIR, exist_ok=True)
-
-# Map user email to safe filename
-def get_user_history_path(email: str) -> str:
-    safe_email = email.replace("@", "_at_").replace(".", "_dot_")
-    return os.path.join(CHAT_HISTORY_DIR, f"{safe_email}.json")
-
-# Load chat history
-def load_history(email: str):
-    path = get_user_history_path(email)
-    try:
-        if os.path.exists(path):
-            with open(path, "r") as f:
-                return json.load(f)
-    except json.JSONDecodeError:
-        logger.info("chat history for %s is corrupted. starting fresh.", email)
-    return []
-
-# Limit history size for memory efficiency
-MAX_HISTORY_LENGTH = 20
-
-# Save chat history
-def save_history(email: str, history):
-    trimmed_history = history[-MAX_HISTORY_LENGTH:]
-    path = get_user_history_path(email)
-    with open(path, "w") as f:
-        json.dump(trimmed_history, f, indent=2)
-
-# Wrapper for asking Gemini model
-def ask_gemini(message: ZulipMessage, model_name: str, prompt: str, use_history: bool):
-    text_error = False
-    try:
-        model = genai.GenerativeModel(
-            model_name=model_name,
-            system_instruction=prompt
-        )
-
-        if use_history: # Chat mode with history
-            history = load_history(message.sender_email)
-            chat_session = model.start_chat(history=history)
-            reply = chat_session.send_message(message.content)
-        else: # One-shot request
-            reply = model.generate_content(message.content)
-
-        if reply.candidates and reply.candidates[0].content.parts:
-            return reply
-        else:
-            text_error = True
-            raise Exception("no text returned")
-
-    except Exception as e:
-        logger.error("gemini model %s failed", model_name, exc_info=True)
-        if text_error:
-            return ask_gemini(message, "gemini-2.5-pro", prompt, use_history)
-        elif model_name == "gemini-2.5-pro":
-            return ask_gemini(message, "gemini-2.5-flash", prompt, use_history)
-        else:
-            raise HTTPException(status_code=500, detail="Gemini model failed")
-
-
-
-# Main logic for handling user messages
-def process_user_message(message):
-    try:
-        history = load_history(message.sender_email)
-
-        logger.info("%s sent message '%s'", message.sender_email, message.content)
-
-        # Save user message to history
-        history.append({"role": "user", "parts": [message.content]})
-        save_history(message.sender_email, history)
-
-        # Determine intent
-        intent_response = ask_gemini(message, current_model, intent_prompt, use_history=False)
-        intent = intent_response.text.strip().lower()
-
-        logger.info("Intent classified as: %s", intent)
-
-        response = ""
-
-        # If user asked about the chatbot itself
-        if intent == "chatbot":
-            chatbot_reply = ask_gemini(message, current_model, chatbot_prompt, use_history=True)
-            history.append({"role": "model", "parts": [chatbot_reply.text]})
-            save_history(message.sender_email, history)
-
-            response = chatbot_reply.text
-
-        # If user asked an unsupported question
-        elif intent == "other":
-            other_reply = ask_gemini(message, current_model, other_prompt, use_history=True)
-            history.append({"role": "model", "parts": [other_reply.text]})
-            save_history(message.sender_email, history)
-
-            response = other_reply.text
-
-        # If user wants to query database
-        elif intent == "database":
-            sql_message = ask_gemini(message, current_model, sql_prompt, use_history=True)
-
-            logger.info("SQL generated: %s", sql_message.text)
-
-            history.append({"role": "model", "parts": [sql_message.text]})
-
-            # Ensure query is safe
-            if not is_safe_sql(sql_message.text):
-                raise ValueError("Unsafe SQL query detected — blocked from execution.")
-
-            # Run query
-            database_data = submit_query(sql_message.text)
-
-            # Salvage message if invalid sql generated
-            if database_data != "salvage":
-
-                history.append({"role": "model", "parts": [database_data]})
-
-                answer_content = (
-                    f"The SQL query returned {database_data}. "
-                    "Answer the user's question using this data."
-                )
-            else:
-                answer_content = (
-                    "A different message instead of SQL was generated. Use this"
-                    f"message to answer the user's question. Message: {database_data}"
-                )
-
-            # Ask Gemini to turn raw SQL results into natural-language answer
-            answer_request = ZulipMessage(
-                content=answer_content,
-                display_recipient=message.display_recipient,
-                sender_email=message.sender_email,
-                subject=message.subject,
-                type=message.type
-            )
-
-            answer_message = ask_gemini(answer_request, current_model, answer_prompt, use_history=True)
-
-            history.append({"role": "model", "parts": [answer_message.text]})
-
-            save_history(message.sender_email, history)
-
-            response = answer_message.text
-
-        logger.info("Fred response: %s", response)
-
-        # Send response back to Zulip user
-        send_zulip_message(
-            to=[message.sender_email],
-            msg_type=message.type,
-            subject=message.subject,
-            content=response,
-            channel_name = message.display_recipient
-        )
-
-    except Exception as e:
-        logger.error("Error: %s", e)
-
-# FastAPI endpoint to handle incoming Zulip messages
-@app.post("/chat")
-def chat(request: ChatRequest, background_tasks: BackgroundTasks):
-    # Verify authentication
-    if request.token != zulip_auth_token:
-        raise HTTPException(status_code=401, detail = "Unauthorized Request")
-
-    # Add user message to background process
-    thinking_reply = "One moment, generating response..."
-    background_tasks.add_task(process_user_message, request.message)
-
-    # Send immediate temp reply
-    try:
-        send_zulip_message(
-            to=[request.message.sender_email],
-            msg_type=request.message.type,
-            subject=request.message.subject,  # Stream messages need subject
-            content=thinking_reply,
-            channel_name=request.message.display_recipient
-        )
-
-    except Exception as e:
-        logger.error("Send Zulip Message Failed", exc_info=True)
-        raise HTTPException(status_code=500, detail = "Send Zulip Message Failed")
-
-    # Return response to Zulip server (not the user)
-    return JSONResponse(content={"response_not_required": True}, status_code=200)
-
+    uvicorn.run("main:app", host="0.0.0.0", port=8000, reload=True)
