+++ pyproject.toml
@@ -22,3 +22,11 @@ disallow_untyped_defs = true
 [tool.pytest.ini_options]
 addopts = "-q --cov=fred_zulip_bot --cov-report=term-missing"
 testpaths = ["tests"]
+
+[[tool.mypy.overrides]]
+module = ["requests", "google.generativeai", "mysql.connector"]
+ignore_missing_imports = true
+
+[[tool.mypy.overrides]]
+module = ["config"]
+ignore_errors = true
